requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
requests==2.31.0
Werkzeug==3.0.1
email-validator==2.1.0.post1
openai==0.28.0
google-generativeai==0.3.0
gunicorn==20.1.0 